feat: Page Component 조합
